검색 및 내브바 컴포넌트화
필요한 페이지에서 적절하게 사용할 수 있게 검색 및 내브바 컴포넌트화
