+++ assets/css/pretendard.css
@@ -1,14 +1,3 @@
-/*
-IE6-IE8
-url('../fonts/test.woff2') format('woff2')
-
-Modern Browsers
-url('../fonts/test.woff') format('woff')
-
-Safari, Android, iOS
-url('../fonts/test.ttf') format('truetype'),
-*/
-
 @font-face {
 	font-family: 'Pretendard';
 	font-weight: 900;
