+++ src/components/topbar/TopNavBar.jsx
@@ -4,39 +4,10 @@ import facepaint from 'facepaint'
 
 import { useMediaQuery } from "react-responsive"
 
-import { Link } from 'react-router-dom';
-
 const mq = facepaint([
   '@media (min-width: 1024px)',
   '@media (min-width: 1440px)',
 ])
-
-const Top = styled.div(() => mq({
-  marginLeft: ['15.5vw', '18.75rem', '18.75rem'],
-  position: 'fixed',
-  left: 0,
-  right: 0,
-  top: 0,
-  zIndex: 1,
-}));
-
-const Top_SearchContainer = styled.div(() => mq({
-  width: '100%',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'flex-end',
-  backgroundColor: '#F4F4F4',
-  paddingTop: ['6%', '1.25rem', '1.25rem'],
-  paddingRight: ['6%', '1.25rem', '1.25rem'],
-  '& p': {
-    marginRight: ['4.6vw', '0.875rem', '0.875rem'],
-    fontSize: ['4.6vw', '0.875rem', '0.875rem'],
-    color: '#828282',
-  },
-  '& img': {
-    width: ['10vw', '2.5rem', '2.5rem'],
-  },
-}))
 
 const Top_NavBar = styled.div(() => mq({
   width: '100%',
@@ -64,15 +35,7 @@ export default function TopNavBar({ pointFont }) {
   });
 
   return (
-    <Top>
-      <Link to='/search'>
-        <Top_SearchContainer>
-          <p>지역, 음식 또는 가게이름</p>
-          <img
-            src='https://img-s3-bucket.s3.ap-northeast-2.amazonaws.com/icon/search-icon.svg'
-          />
-        </Top_SearchContainer>
-      </Link>
+    <>
       <Top_NavBar>
         <p
           className={
@@ -116,6 +79,6 @@ export default function TopNavBar({ pointFont }) {
           {isPc ? '나에게 딱 맞는 맞집 추천!' : '맛집 결과!'}
         </p>
       </Top_NavBar>
-    </Top>
+    </>
   )
 }
